PLATUI-1141: update backstop config to increase test coverage
Backstop is used for visual regression testing, previously we only compared the
default example for each component. This change updates the config so that all
examples are included in the backstop test coverage. I have chosen to separate
the commit that will update the backstop reference data to make this change
easier to review.