NOJIRA: fix header language select underline
I've updated css for language select, but as part of this I had to
increase the sensitivity of backstop to pick up the fix. However that
highlighted that there are lots of other cases where the sensitivity has
been too low to pick up changes in big screenshots. So I'm leaving the
sensitivity settings as they are here and hopefully that doesn't cause
any false positives.