Patch some bits raised by npm audit (#211)
* PLATUI-1623: Resolve warnings from npm audit

By updating these dependencies:
- update backstopjs from v5 to v6
- update node-sass@7.0.1
- update standard

* NOJIRA: fix header language select underline

I've updated css for language select, but as part of this I had to
increase the sensitivity of backstop to pick up the fix. However that
highlighted that there are lots of other cases where the sensitivity has
been too low to pick up changes in big screenshots. So I'm leaving the
sensitivity settings as they are here and hopefully that doesn't cause
any false positives.

* NOJIRA: revert bug I added a long time ago

Apparently at some point I set the diff threshold on internal-header to
1% because I thought a change was a rendering difference between
platforms but actually it was (I guess) that the padding or margin
changed. Spent a long time trying to debug this, because I couldn't
understand why backstop was not failing internal-header screenshots.