PLATUI-1691: install better-npm-audit (#227)
And setup some exclusions for current issues with devDependencies that
we've checked and don't seem to be a risk with our usage. For most there's no
updated version we can use of our direct dependencies that addresses the
issue, and patching the transitive dependencies can break the usage or
introduce it's own security issues.