PLATUI-2106: implement failing tests for the bug
Had to update puppeteer and tweak account-menu test as a result

PLATUI-2106: fix bug with timeout-dialog

It could cause an early timeout because the currentTimeout that starts
when a timeout dialog is opened would only be cleaned up properly the
first time the timeout dialog was opened. Exacerbated by the tab sync
code which could cause it to trigger even earlier.

Co-authored-by: Katie Pocklington <katie.pocklington@digital.hmrc.gov.uk>