PLATUI-2847: revert back to previous structure
In trying to create an example of the issue that I changed the structure
to fix, I realised that the issue came not from a govuk change but from
changing back to follow the govuk convention. Because of that, even
though I still think it would be good to match GOV.UK - it's not worth
making that change as part of this - especially because we still have
the language select and sign out links within the header which mean we
wouldn't be able to get rid of our fork all together.