PLATUI-3037: preload fonts when previewing to avoid FOUT
FOUT is Flash Of Unstyled Text

the timeout dialog tests were flakey because no other text using the
govuk font was loaded before the dialog opened, so because browsers wait
to load fonts until they are needed, when the dialog opened there was
chance for a race where screenshots for VRTs are captured before the
fonts have finished loading.

as well as preloading, we could get puppeteer to wait for
document.fonts.ready - however it might not be needed to not implemented
here at the moment.