PLATUI-3037: suppress sass warnings from our styles in prototypes (#373)
* PLATUI-3037: suppress sass warnings from our styles in prototypes

One I could fix, the others we need to suppress - and the reasons have
been added as comments inline - off the back of this we need to talk
with pta and design resources about change needed to pta-account-menu

* PLATUI-3037: fix a very rarely flakey test

the suffix in the timeout dialog text filled the width of the timeout
dialog so perfectly that very occaisionally it could wrap to another
line if the calculations rounded out slightly differently. So we have
shortened the suffix.

* PLATUI-3037: preload fonts when previewing to avoid FOUT

FOUT is Flash Of Unstyled Text

the timeout dialog tests were flakey because no other text using the
govuk font was loaded before the dialog opened, so because browsers wait
to load fonts until they are needed, when the dialog opened there was
chance for a race where screenshots for VRTs are captured before the
fonts have finished loading.

as well as preloading, we could get puppeteer to wait for
document.fonts.ready - however it might not be needed to not implemented
here at the moment.