Proof of concept: migrate to dart-sass
The key bit of this proof of concept is that by bodging the includePaths
we've made it so that we get the behaviour we had before from libsass
which means we don't need to change the relative imports we have in our
sass paths that will require sass load paths be updated by consumers.
Where some of our consumers are using different ways of loading packages
and they won't always be in node_modules folder, and some don't have the
option of updating their load paths (govuk prototype kit) users.