PLATUI-3901: adding in some more explicit waits to combat vrt test flakiness (#478)
haven't really been able to find a way to put my finger on the scale such that the
flakiness we're experiencing becomes more reproducible so opting instead to just
add in some explicit timeouts after actions we're experiencing flakiness with in a
bit of a brute force attempt. I have omitted npm version bump to avoid a release
with no functional changes.